backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
passlib[argon2,bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
python-dotenv==1.0.1
resend==0.7.0
